Hotplug fix possible Unhandled Error
We always ensure hotplug.socket in tmp for twisted/internet/unix.py in connectionLost

+++ lib/python/Plugins/SystemPlugins/Hotplug/plugin.py
@@ -87,6 +87,7 @@ def autostart(reason, **kwargs):
 	if reason == 0:
 		from twisted.internet import reactor
 		try:
+		   if os.path.exists("/tmp/hotplug.socket"):
 			os.remove("/tmp/hotplug.socket")
 		except OSError:
 			pass
